refactor : 상수명은 컨벤션에 맞게 작성, 가독성을 위해 String.format() 활용

+++ src/main/java/lotto/domain/LottoRankRecord.java
@@ -5,7 +5,7 @@ import java.util.Objects;
 public class LottoRankRecord {
 
     private final LottoRank lottoRank;
-    private final long number;
+    private final int number;
 
     public LottoRankRecord(LottoRank lottoRank, int number) {
         this.lottoRank = lottoRank;
